--- a/calculators/sacksCalculator.js
+++ b/calculators/sacksCalculator.js
@@ -1,13 +1,13 @@
-const skyblockItems = require('../constants/items.json');
 const { validRunes } = require('../constants/misc');
 const { titleCase } = require('../helper/functions');
+const { getHypixelItemInformationFromId } = require("../constants/itemsMap")
 
 const calculateSackItem = (item, prices) => {
   const itemPrice = prices[item.id.toLowerCase()] || 0;
   if (item.id.startsWith('RUNE_') && !validRunes.includes(item.id)) return null;
   if (itemPrice) {
     return {
-      name: item.name || skyblockItems.find((skyblockItem) => skyblockItem.id === item.id)?.name || titleCase(item.id),
+      name: item.name || getHypixelItemInformationFromId(item.id)?.name || titleCase(item.id),
       id: item.id,
       price: itemPrice * item.amount,
       calculation: [],