Adds itemsMap.js into constants folder (lmk if I should move it to helpers folder) that stores a map from item id -> item data
So that you dont need to loop over the whole items array every time you want to search for that data

In my testing it brings the networth loading time from ~300ms to ~50ms
I did get it to take 5ms per but I cant replicate it (also not sure how cus it takes ~40ms for decoding the items and ~10ms for atual nw)