@MattTheCuber 2022-12-08
Wouldn't it make sense to create a itemsMap.json and maybe a itemsMap.js that requests the /items endpoint, parses, and writes to itemsMap.json that way the first check doesn't take an extra 250ms?